% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\docType{data}
\name{spikes}
\alias{spikes}
\title{Neural spike data for mini lab I, BCS 511}
\format{An object of class \code{"data.frame"} with 45,841 rows and 14 variables:
\describe{
  \item{ID}{ID identifying unique Animal, Cell, and Run.}
  \item{Animal}{Identifier for the Animal recorded from (2 levels: m1 or m2).}
  \item{Cell}{Identifier for the cell recorded from (85 levels).}
  \item{Run}{Identifier for the recording run (7 levels).}
  \item{Trial}{Trial on which the stimulus was shown and the neural spikes and choice was collected.}
  \item{DisparityFar}{Whether the stimulus disparity was far (1), near (-1) or not defined (0: for binocular correlations of 0\%).}
  \item{DisparityPreferred}{Whether the stimulus disparity was the preferred one for the cell (1), null/anti-preferred (-1), or not defined (0: for binocular correlations of 0\%).}
  \item{BinocularR}{Binocular correlation of stimulus (in \%). Multiple levels tested for each cell, but the levels tested can differ between cells.}
  \item{Spikes}{Spike rate (spikes/time) observed in the cell during exposure to the stimulus.}
  \item{Choice}{Animal's response at the end of the trial (1 for far; -1 for near).}
  \item{Near}{Cell's spike rate for near stimulus during disparity tuning.}
  \item{Far}{Cell's spike rate for far stimulus during disparity tuning.}
  \item{Correct}{Whether animal's response was correct (1) or incorrect (0).}
}}
\source{
\href{http://www.sas.rochester.edu/bcs/people/faculty/deangelis_greg/index.html}{DeAngelis lab}.
Prepared by Marjena Popovic. Shared for BCS 511 with permission. Do not distribute further.
}
\usage{
data(spikes)
}
\description{
Neural spike data for mini lab I, BCS 511
}
\references{
Uka and DeAngelis (2003) The Journal of Neuroscience, 23(8):3515–3530.
}
\keyword{datasets}
